Add class to extract metadata from MRC files
Often our input files contain metadata that would be
useful to include in the mmCIF file. Start adding
classes to extract this metadata from each
domain-specific file type. For example, MRC files
used for EM maps may contain an EMDB identifier,
which we can point to in preference to the local file.