Add method to point Locations to Repositories
If the current working directory is a checkout of
a repository which is archived somewhere with a URL,
it is helpful to construct simple local file references
(FileLocation objects with repo=None) and then update
them all at the end of the script to point to the
repository. This makes it easy to work with local files
and switch to archived copies at a DOI later.