Handle multi-chain PDB file templates
If a PDB file is a comparative model and it contains multiple
chains, the returned Template objects previously did not identify
which chain they modeled. Modify the metadata parser to return
separate lists of Templates for each chain, so we can pick the
chain we're actually interested in (e.g. for a StartingModel).