Use namedtuple rather than dict
Rather than passing a dict of values to each
category handler, pass a namedtuple instead.
This should be more efficient, in the usual
case where the keys we extract are a subset of
the keys in the file, since this avoids allocating
memory for the dict keys, which never change.