Start handling poly_seq_scheme entirely in C
Since the pdbx_poly_seq_scheme table is typically
large and contains little information (often it is
a 1-to-1 mapping) we can speed up reading an mmCIF
file by handling this table entirely in C, skipping
many Python method calls. This provides an initial
implementation which stores the information in C
data structures, but does not populate the final
Python objects yet.