Add mostly complete write support for BinaryCIF
BinaryCIF files can now be written in the same
fashion as mmCIF, by specifying the 'format'
argument to ihm.dumper.write(). These files should
handle all of the ihm categories, and should be
valid BinaryCIF format, but they are not as compact
as could be - in particular no delta or run length
encoding is yet done on integer data.