Allow Dumper.finalize() to be called multiple times
Previously we assumed that Dumper.finalize() was
only ever called once, and so assumed that before
calling util._assign_id() no objects had IDs. This
assumption does not hold if the user dumps a System
multiple times (e.g. in both mmCIF and BinaryCIF format).
Clear object IDs at the start of finalize() so that
they are reassigned correctly.