Add an 'end save frame' category callback
When handling a save frame, it is helpful for the handlers
to know when the reader reached the end of the frame so
that they can store the final data. To facilitate this,
and an 'end save frame' callback, similar to the existing
finalize callback (which is called at the end of the data
block). In Python this simply calls the _Handler.end_save_frame
method.