Allow features selecting entire nonpoly asyms
Update to match the latest IHM dictionary, renaming
the ihm_non_poly_atom_feature category to
ihm_non_poly_feature. The dictionary allows for an
entire nonpolymeric asym unit to be selected (by
specifying '.' for the atom id) so generalize the
Feature class to take a wider variety of objects,
for both polymers and nonpolymers.