Representation need not be a subset of assembly
Relax our checks which required that a model's
Representation only ever represented segments that
were also in the model's Assembly. (A representation
can include segments outside of the assembly when
performing multi-state modeling.) Instead, only
require that any Atoms or Spheres emitted for a
model are within both the representation and the
assembly. Closes #27.