Only query EMDB API when needed
Rather than querying the EMDB web API when we parse
the MRC file, wait until the EMDB-derived information
(version & details) is actually asked for. This avoids
hitting the network at all if the information is never
needed, and also allows for the EMDB accession code to
be manually reset if necessary. Closes #34.