Don't read files by fd under Python 3
PyObject_AsFileDescriptor() returns a valid descriptor
for some filelike objects that does not correspond to
the file we want to read (e.g. those opened with
gzip.open() refer to the compressed file, not the
uncompressed data) so we can't trust it under Python 3 -
always read data via the read() method.