Handle non-ASCII mmCIF files with C parser
If given an mmCIF file in a non-ASCII encoding
(e.g. UTF-8) the C parser would previously fail
with "Python read method returned too many bytes"
because a result in *characters* was used to fill
a buffer in *bytes* and many Unicode characters
need more than 1 byte to be represented in UTF-8.
Closes #53.