Do not give assembly description for protocol
Do not fill out the struct_assembly_description item
in the ihm_modeling_protocol_details table using
Assembly.description; just omit it. At best providing
this just duplicates information (since we already provide
the assembly ID) and at worst it violates the mmCIF format
(since the type for this data item is not the same as
that for the assembly description). Closes #63.