Allow for custom CCDs for ChemComps
Provide support for describing chemical components
that are not in the wwPDB CCD by adding additional
arguments to the ChemComp constructor. The IHM
dictionary does not support this, so raise an error
if a user tries to use these in an IHM file. This is
primarily for use by python-modelcif, as the MA
dictionary *does* provide support.
Relates ihmwg/python-modelcif#21.