Quote strings starting with reserved words
Currently we quote strings when they match a
reserved word such as "stop_". The STAR format
is ambiguous over whether a string that only
starts with a reserved word, such as "stop_at_score",
needs to be quoted, but some tools, such as the
GEMMI library, have trouble reading files where
these strings are not quoted. To be on the safe
side, quote all such strings.
Closes ihmwg/python-modelcif#25.