Attempt to disambiguate water numbering
Require that any water Atom be given a seq_id
(rather than None) so that it is unambiguous,
and so that the dumper can write the correct
auth_seq_id for it in the atom_site table.
Since there can only be one water Entity but
potentially multiple asyms (of potentially
different lengths), move the specification of
number of waters (and thus the sequence and seq_id_range)
from Entity to a new WaterAsymUnit class, which
must be used rather than AsymUnit for waters.
Relates #98, but does not completely resolve
the issue of multiple waters as we still don't
handle reading correctly (we will need to reverse
map auth_seq_id to our internal seq_id).