Allow branched entities with same "sequence"
It is quite possible that a system may contain
multiple branched entities with the same "sequence"
but different connectivity, so relax our "entities
cannot have the same sequence" check for branched
entities. To be strictly correct, we should check
for branched entities that have both the same
sequence *and* the same connectivity, but that is not
handled here. Closes #126.