Pass through the pdbx_database_status category
If the pdbx_database_status category is present
in an input file, read it in as a Python dict,
so that we can write it out again more or less
unchanged. This allows the make_mmcif utility to
work on deposited files while preserving this data.
Store the resulting data in the (undocumented,
and subject to change) System._database_status member.
Closes #135.