Check residue ranges at creation time
Make sure that any EntityRange or AsymUnitRange
only references valid seq_ids, when we create
the objects. This check can be disabled with
an undocumented _check parameter, which we use
when reading files, as the files may contain
invalid ranges (or the full Entity sequence may not
have been read yet when we read a range).
Even if the check is skipped, however, the range
will still be checked at dump time.