Be more flexible when comparing Assemblies
Provide an Assembly._signature() method which
returns a Python object that represents the
Assembly and that (unlike the Assembly itself)
is hashable. Use this to compare Assemblies.
This makes our detection of "equal" Assemblies
more flexible so that order and domain boundaries
are ignored. For example, the three assemblies
(A(1-100), B); (B, A(1-100));
and (A(1-50), A(51-100), B)
which were previously considered different,
are now considered equal. Closes #170.